% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/horseshoe.R
\name{compute_horseshoe_loo}
\alias{compute_horseshoe_loo}
\title{Compute LOO-CV for Horseshoe Model}
\usage{
compute_horseshoe_loo(fit, verbose = FALSE)
}
\description{
Compute LOO-CV for Horseshoe Model
}
\keyword{internal}
