% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/horseshoe.R
\name{evaluate_fit_quality_internal}
\alias{evaluate_fit_quality_internal}
\title{Evaluate Fit Quality}
\usage{
evaluate_fit_quality_internal(
  n_divergences,
  max_treedepth_hit,
  low_bfmi,
  rhat_max,
  ess_bulk_min,
  ess_tail_min
)
}
\description{
Evaluate Fit Quality
}
\keyword{internal}
