% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/horseshoe.R
\name{extract_mcmc_diagnostics}
\alias{extract_mcmc_diagnostics}
\title{Extract MCMC Diagnostics}
\usage{
extract_mcmc_diagnostics(fit, verbose = FALSE)
}
\arguments{
\item{fit}{cmdstanr fit object.}

\item{verbose}{Logical for messages.}
}
\value{
List of diagnostic metrics.
}
\description{
Extracts comprehensive MCMC diagnostics from a cmdstanr fit object.
}
\keyword{internal}
