% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal_analysis.R
\name{generate_interpretation}
\alias{generate_interpretation}
\title{Generate Automated Technical Interpretation}
\usage{
generate_interpretation(results, verbose = FALSE)
}
\arguments{
\item{results}{signal_analysis results object}

\item{verbose}{Logical, print progress}
}
\value{
List containing interpretation components
}
\description{
Internal function to synthesize results into interpretable narrative.
}
\keyword{internal}
