% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/horseshoe.R
\name{horseshoe}
\alias{horseshoe}
\title{Regularized Horseshoe Regression for Variable Selection}
\description{
Implements Bayesian sparse regression using the regularized Horseshoe prior
for identifying structurally relevant predictors from high-dimensional
candidate variable sets.
}
