% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal_analysis.R
\name{iplot}
\alias{iplot}
\title{Interactive Plot for Signal Analysis}
\usage{
iplot(x)
}
\arguments{
\item{x}{An object of class \code{signal_analysis}}
}
\value{
A plotly object (HTML widget).
}
\description{
Generates an interactive dashboard using plotly to explore the results of
the signal analysis. Allows zooming, panning, and toggling traces.
}
