% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/horseshoe.R
\name{posterior_predictive_check_horseshoe}
\alias{posterior_predictive_check_horseshoe}
\title{Posterior Predictive Check for Horseshoe Model}
\usage{
posterior_predictive_check_horseshoe(fit, y, verbose = FALSE)
}
\description{
Posterior Predictive Check for Horseshoe Model
}
\keyword{internal}
