% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal_analysis.R
\name{print.signal_analysis}
\alias{print.signal_analysis}
\title{Print Method for signal_analysis Objects}
\usage{
\method{print}{signal_analysis}(x, ...)
}
\arguments{
\item{x}{An object of class \code{signal_analysis}}

\item{...}{Additional arguments (ignored)}
}
\value{
Invisibly returns the input object
}
\description{
Print a concise summary of signal analysis results.
}
