% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/horseshoe.R
\name{process_horseshoe_results}
\alias{process_horseshoe_results}
\title{Process Horseshoe Results}
\usage{
process_horseshoe_results(fit, var_names, verbose = FALSE)
}
\description{
Process Horseshoe Results
}
\keyword{internal}
