% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_dfm.R
\name{procrustes_rotation}
\alias{procrustes_rotation}
\title{Procrustes Rotation for Bootstrap Alignment}
\usage{
procrustes_rotation(B, A)
}
\arguments{
\item{B}{Matrix to rotate.}

\item{A}{Target matrix.}
}
\value{
Rotated matrix B.
}
\description{
Rotates matrix B to best match target matrix A using orthogonal Procrustes
rotation.
}
\keyword{internal}
