% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_root.R
\name{synthesize_unitroot_results}
\alias{synthesize_unitroot_results}
\title{Synthesize Unit Root Results}
\usage{
synthesize_unitroot_results(results, alpha)
}
\description{
Synthesize Unit Root Results
}
\keyword{internal}
