% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/horseshoe.R
\name{get_horseshoe_stan_code}
\alias{get_horseshoe_stan_code}
\title{Get Stan Code for Regularized Horseshoe}
\usage{
get_horseshoe_stan_code()
}
\value{
Character string containing Stan code.
}
\description{
Returns the Stan code for the regularized Horseshoe model. This is used
when the package is not installed and the inst/stan file is not available.
}
\keyword{internal}
